update peekbuffer api, convergence with official arduino api

+++ cores/esp8266/FS.h
@@ -63,7 +63,7 @@ public:
     size_t readBytes(char *buffer, size_t length)  override {
         return read((uint8_t*)buffer, length);
     }
-    size_t read(uint8_t* buf, size_t size);
+    int read(uint8_t* buf, size_t size);
     bool seek(uint32_t pos, SeekMode mode);
     bool seek(uint32_t pos) {
         return seek(pos, SeekSet);
